requirements.txt
aiohttp==3.7.4.post0
alpaca-trade-api==1.2.1
async-timeout==3.0.1
attrs==21.2.0
cbpro==1.1.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
dateparser==1.0.0
Flask==2.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
msgpack==1.0.2
multidict==5.1.0
numpy==1.20.3
pandas==1.2.4
pymongo==3.5.1
python-binance==1.0.10
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.13.0
six==1.10.0
sortedcontainers==2.3.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websocket-client==0.40.0
websockets==8.1
Werkzeug==2.0.0
yarl==1.6.3
